Rename package name to commonmark
This makes the package name PEP8-compliant by changing:

    import CommonMark

to:

    import commonmark

The old camel-cased import should still work as well though, for now.

Closes #60
