--- CommonMark/__init__.py
@@ -1,8 +1,0 @@
-# flake8: noqa
-from __future__ import unicode_literals, absolute_import
-
-from CommonMark.main import commonmark
-from CommonMark.dump import dumpAST, dumpJSON
-from CommonMark.blocks import Parser
-from CommonMark.render.html import HtmlRenderer
-from CommonMark.render.rst import ReStructuredTextRenderer
