+++ MANIFEST.in
@@ -2,17 +2,17 @@ include README.rst
 include LICENSE
 include .gitignore
 include spec.txt
-include CommonMark/__init__.py
-include CommonMark/blocks.py
-include CommonMark/common.py
-include CommonMark/dump.py
-include CommonMark/entitytrans.py
-include CommonMark/inlines.py
-include CommonMark/main.py
-include CommonMark/node.py
-include CommonMark/utils.py
-include CommonMark/render/__init__.py
-include CommonMark/render/renderer.py
-include CommonMark/render/html.py
-include CommonMark/tests/run_spec_tests.py
-include CommonMark/tests/unit_tests.py
+include commonmark/__init__.py
+include commonmark/blocks.py
+include commonmark/common.py
+include commonmark/dump.py
+include commonmark/entitytrans.py
+include commonmark/inlines.py
+include commonmark/main.py
+include commonmark/node.py
+include commonmark/utils.py
+include commonmark/render/__init__.py
+include commonmark/render/renderer.py
+include commonmark/render/html.py
+include commonmark/tests/run_spec_tests.py
+include commonmark/tests/unit_tests.py
