+++ tox.ini
@@ -9,5 +9,5 @@ envlist = py{26,27,33,34,35,36,py,py3}
 [testenv]
 deps = .[test]
 commands =
-    python -m CommonMark.tests.unit_tests
-    python -m CommonMark.tests.run_spec_tests {posargs:}
+    python -m commonmark.tests.unit_tests
+    python -m commonmark.tests.run_spec_tests {posargs:}
